Implement table tennis league
Create an app which organises a pyramid-shaped table-tennis ladder.

Players are added to the league so that one person is in the top rung,
two people in the second, three people in the third and so on. The
person at the top is the current winner.

Players can callenge a player in the rung above them (but no higher). If
the lower player wins, the players swap places.